Fix createMemo & indexArray to work with functors
As createMemo and indexArray may accept any kind of return element for there memo and iterator functions,
it is possible for a developer to provide a functor (a function which returns a function).

We need to account for this possibility when using the returned value in a signal, since signal setters will treat functions as setter function (prev: T) => T
instead of recognizing the function itself as the next value for the signal.